requirements.txt
# Core dependencies
numpy>=1.21.0,<2.0
pandas>=1.3.0,<3.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0

# Data processing
pyyaml>=6.0
requests>=2.26.0

# Modeling
xgboost>=1.5.0
lightgbm>=3.3.0
joblib>=1.3.0
shap>=0.44.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# Streamlit
streamlit>=1.36,<2.0

# Development (optional)
pytest>=6.2.0
black>=21.12b0
flake8>=4.0.0
jupyter>=1.0.0
